requirements.txt
# MCP智能路由系统依赖包
# 核心依赖
mcp>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
pymysql>=1.1.0

# 网络请求
requests>=2.31.0

# 异步支持
asyncio-mqtt>=0.16.0

# 开发和测试依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# 可选依赖（根据需要安装）
# Jupyter支持
# jupyter>=1.0.0
# ipykernel>=6.0.0

# 数据库支持
# sqlalchemy>=2.0.0
# alembic>=1.12.0